Результат тип поля "файл" - перенос типов + новый компонент
