+++ l2-frontend/src/construct/ParaclinicResearchEditorComponents/FieldKinds/FileField.vue
@@ -171,8 +171,9 @@ import {
 import Treeselect from '@riophae/vue-treeselect';
 import '@riophae/vue-treeselect/dist/vue-treeselect.css';
 
-import { FileFieldConstructorSettings, FileFieldSettings, SelectOption } from
+import { FileFieldConstructorSettings, SelectOption } from
   '@/construct/ParaclinicResearchEditorComponents/types/FileField';
+import { FileFieldSettings } from '@/types/Descriptive/Fields/FileField';
 
 const props = defineProps<{
   value: FileFieldSettings | null
