+++ l2-frontend/src/construct/ParaclinicResearchEditorComponents/types/FileField.ts
@@ -1,30 +1,3 @@
-export type FileRulesMode = 'exact' | 'oneOf'
-
-export type FileRuleItem = {
-  extension: string
-  count: number
-}
-
-export type FileRuleVariant = {
-  items: FileRuleItem[]
-}
-
-export type FileFieldSettings = {
-  minFiles: number | null
-  maxFiles: number | null
-  maxFileSizeMb: number | null
-  maxTotalSizeMb: number | null
-
-  allowedExtensions: string[]
-
-  filenamePattern: string
-  filenamePatternDescription: string
-  strictFilename: boolean
-
-  rulesEnabled: boolean
-  rulesMode: FileRulesMode
-  rulesVariants: FileRuleVariant[]
-}
 export type SelectOption = {
   id: string | number
   label: string
